+++ source/extensions/lightspeed.event.capture_persp_to_persp/lightspeed/event/capture_persp_to_persp/tests/unit/test_core.py
@@ -30,7 +30,7 @@ from lightspeed.layer_manager.core import LayerType as _LayerType
 from lightspeed.trex.capture.core.shared import Setup as _CaptureCoreSetup
 from lightspeed.trex.viewports.shared.widget import create_instance as _create_viewport_instance
 from omni.flux.utils.widget.resources import get_test_data as _get_test_data
-from omni.kit.test.async_unittest import AsyncTestCase
+from omni.kit.test import AsyncTestCase
 from omni.kit.test_suite.helpers import arrange_windows, open_stage, wait_stage_loading
 
 # from omni.kit.test_suite.helpers wait_stage_loading