+++ source/extensions/lightspeed.trex.properties_pane.shared.asset_replacements.widget/lightspeed/trex/properties_pane/shared/asset_replacements/widget/tests/e2e/test_widget.py
@@ -26,7 +26,7 @@ from lightspeed.trex.properties_pane.shared.asset_replacements.widget import (
 )
 from omni.flux.utils.widget.resources import get_test_data as _get_test_data
 from omni.kit import ui_test
-from omni.kit.test.async_unittest import AsyncTestCase
+from omni.kit.test import AsyncTestCase
 from omni.kit.test_suite.helpers import arrange_windows, open_stage, wait_stage_loading