+++ source/extensions/lightspeed.trex.selection_tree.shared.widget/lightspeed/trex/selection_tree/shared/widget/tests/e2e/test_widget.py
@@ -34,7 +34,7 @@ from omni.flux.utils.common import path_utils as _path_utils
 from omni.flux.utils.widget.resources import get_test_data as _get_test_data
 from omni.flux.validator.factory import BASE_HASH_KEY
 from omni.kit import ui_test
-from omni.kit.test.async_unittest import AsyncTestCase
+from omni.kit.test import AsyncTestCase
 from omni.kit.test_suite.helpers import arrange_windows, open_stage, wait_stage_loading