+++ source/extensions/lightspeed.trex.viewports.manipulators/lightspeed/trex/viewports/manipulators/tests/e2e/test_widget.py
@@ -22,7 +22,7 @@ import omni.usd
 from lightspeed.trex.viewports.shared.widget import create_instance as _create_viewport_instance
 from omni.flux.utils.widget.resources import get_test_data as _get_test_data
 from omni.kit import ui_test
-from omni.kit.test.async_unittest import AsyncTestCase
+from omni.kit.test import AsyncTestCase
 from omni.kit.test_suite.helpers import open_stage, wait_stage_loading
 from omni.kit.ui_test import Vec2
 from pxr import Gf