+++ source/extensions/omni.flux.validator.manager.widget/omni/flux/validator/manager/widget/tests/e2e/test_widget.py
@@ -25,7 +25,7 @@ from omni.flux.validator.manager.core import ManagerCore as _ManagerCore
 from omni.flux.validator.manager.widget import ValidatorManagerWidget as _ValidatorManagerWidget
 from omni.flux.validator.plugin.check.usd.example.print_prims import PrintPrims as _PrintPrims
 from omni.kit import ui_test
-from omni.kit.test.async_unittest import AsyncTestCase
+from omni.kit.test import AsyncTestCase
 from omni.kit.test_suite.helpers import arrange_windows, get_test_data_path, open_stage