+++ source/extensions/omni.flux.validator.mass.queue.widget/omni/flux/validator/mass/queue/widget/tests/e2e/test_widget.py
@@ -19,7 +19,7 @@ import omni.ui as ui
 from omni.flux.validator.mass.queue.widget import MassQueueTreeWidget as _MassQueueTreeWidget
 from omni.flux.validator.mass.queue.widget.service import get_service_instance as _get_service_instance
 from omni.kit import ui_test
-from omni.kit.test.async_unittest import AsyncTestCase
+from omni.kit.test import AsyncTestCase
 from omni.kit.test_suite.helpers import arrange_windows