+++ source/extensions/lightspeed.trex.asset_replacements.core.shared/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [2.3.6]
+## Changed
+- update to use omni.kit.test public api
+
 ## [2.3.5]
 ### Fixed
 - Fixed tests flakiness