+++ source/extensions/omni.flux.layer_tree.usd.widget/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.8.3]
+## Changed
+- update to use omni.kit.test public api
+
 ## [1.8.2]
 ### Fixed
 - Fixed tests flakiness