+++ source/extensions/omni.flux.service.factory/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.3.2]
+## Changed
+- update to use omni.kit.test public api
+
 ## [1.3.1]
 ### Fixed
 - Fixed tests flakiness